pyproject.toml
[project]
name = "job-scrapper"
version = "1.0.0"
description = "Minimal job scraper with notifications"
requires-python = ">=3.10"
dependencies = [
    "python-jobspy>=1.1.0",
    "apprise>=1.7.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
]

[tool.uv]
dev-dependencies = []
